chore(message-system): refactor ab testing to follow code style and make it easier to reuse on mobile
