+++ suite-common/message-system/package.json
@@ -27,6 +27,8 @@
         "fs-extra": "^11.2.0",
         "json-schema-to-typescript": "^13.1.2",
         "jws": "^4.0.0",
+        "react": "18.2.0",
+        "react-redux": "8.0.7",
         "semver": "^7.6.3"
     },
     "devDependencies": {
