+++ suite-common/message-system/src/index.ts
@@ -6,5 +6,5 @@ export * from './messageSystemThunks';
 export * from './messageSystemTypes';
 export * from './messageSystemUtils';
 
-export * from './experiment';
-export * from './experiment/experiments';
+export * from './experimentUtils';
+export * from './useExperiment';
